Make markdown page element move mathjax at build-time
