+++ core/webviz/_page_element.py
@@ -2,7 +2,6 @@ from abc import ABCMeta, abstractmethod
 from uuid import uuid4
 from ._header_element import HeaderElement
 from os import path
-from re import sub
 from ordered_set import OrderedSet
 
 
@@ -58,23 +57,7 @@ class PageElement:
             html = self.get_template().render(element=self)
 
         """
-
-    def add_mathjax(self, src, files):
-        subdir = sub(r'.*node_modules/', '', str(src))
-        for file in files:
-            self.add_resource(path.join(
-                src,
-                file
-            ), subdir='js/' + str(subdir))
-
-        self.header_elements.add(HeaderElement(
-            tag='script',
-            attributes={
-                'src': path.join(
-                    '{root_folder}', 'resources', 'js', 'mathjax', 'MathJax.js'
-                )
-            }
-        ))
+        pass
 
     def __str__(self):
         html = ""
